fix: change cron schedule to daily for Vercel Hobby plan
The send-reminders cron was set to every 6 hours (0 */6 * * *) which
exceeds the Hobby plan limit of once daily. Changed to 07:00 UTC daily.
