+++ vercel.json
@@ -8,7 +8,7 @@
   "crons": [
     {
       "path": "/api/cron/send-reminders",
-      "schedule": "0 */6 * * *"
+      "schedule": "0 7 * * *"
     }
   ]
 }
